# docs/swagger.yaml
definitions:
...
  models.Address:
    properties:
      city:
        type: string
      contact_id:
        type: integer
      country:
        type: string
      extended_street:
        type: string
      id:
        type: integer
      is_primary:
        type: boolean
      label_type_id:
        type: integer
      postal_code:
        type: string
      state:
        type: string
      street:
        type: string
      type_label:
        type: string
    type: object
  models.AddressJSONPatch:
    properties:
      city:
        type: string
      contact_id:
        example: 4
        type: integer
      country:
        type: string
      extended_street:
        type: string
      id:
        example: 1
        type: integer
      is_primary:
        example: false
        type: boolean
      label_type_id:
        example: 42
        type: integer
      postal_code:
        type: string
      state:
        type: string
      street:
        type: string
    type: object
...
  models.ContactDateJSONPatch:
    properties:
      date:
        example: "2026-04-30"
        type: string
      date_day:
        example: 30
        type: integer
      date_month:
        example: 4
        type: integer
      date_type:
        description: '"anniversary", "birthday", or "other"'
        example: anniversary
        type: string
      id:
        example: 123
        type: integer
      other_date_id:
        example: 456
        type: integer
    type: object
...
  models.ContactLabelJSONPost:
    properties:
      category:
        type: string
      name:
        type: string
    type: object
...
  models.OtherDate:
    properties:
      contact_id:
        type: integer
      event_date:
        type: string
      event_date_day:
        description: 1-31, for partial dates
        type: integer
      event_date_month:
        description: 1-12, for partial dates
        type: integer
      event_name:
        type: string
      id:
        type: integer
    type: object
...
info:
  contact:
    name: API Support
    url: https://github.com/steveredden/KindredCard/issues
  description: Personal CRM for managing contacts, relationships, and important dates
  license:
    name: AGPLv3
    url: https://www.gnu.org/licenses/agpl-3.0.txt
  title: KindredCard API
  version: "1.0"
paths:
  /api/v1/addresses/{aid}:
    patch:
      consumes:
      - application/json
      description: Update specific fields of a address using HTTP PATCH. Only provided
        fields will be updated.
      parameters:
      - description: Address ID
        in: path
        name: aid
        required: true
        type: integer
      - description: Address fields to update
        in: body
        name: contact
        required: true
        schema:
          $ref: '#/definitions/models.AddressJSONPatch'
      produces:
      - application/json
      responses:
        "200":
          description: Updated address
          schema:
            items:
              $ref: '#/definitions/models.Address'
            type: array
        "400":
          description: Invalid request body or contact ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Contact not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiTokenAuth: []
      summary: Update a address number
      tags:
      - contacts
...
  /api/v1/contacts/{cid}/addresses:
    post:
      consumes:
      - application/json
      description: Associates a URL with a contact using HTTP POST
      parameters:
      - description: Contact ID
        in: path
        name: id
        required: true
        type: integer
      - description: Address fields
        in: body
        name: url
        required: true
        schema:
          $ref: '#/definitions/models.Address'
      produces:
      - application/json
      responses:
        "200":
          description: 'created: newID'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request body or contact ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Contact not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiTokenAuth: []
      summary: Creates a website / URL associated with a contact
      tags:
      - contacts
  /api/v1/contacts/{cid}/addresses/{a}:
    delete:
      consumes:
      - application/json
      description: Associates a URL with a contact using HTTP POST
      parameters:
      - description: Contact ID
        in: path
        name: cid
        required: true
        type: integer
      - description: URL ID
        in: path
        name: aid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: deleted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request body or contact ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Contact not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiTokenAuth: []
      summary: Creates a website / URL associated with a contact
      tags:
      - contacts
...
  /api/v1/contacts/{cid}/other-dates:
    post:
      consumes:
      - application/json
      description: Associates an Other Date with a contact using HTTP POST
      parameters:
      - description: Contact ID
        in: path
        name: cid
        required: true
        type: integer
      - description: Other Date fields
        in: body
        name: otherDate
        required: true
        schema:
          $ref: '#/definitions/models.OtherDate'
      produces:
      - application/json
      responses:
        "200":
          description: 'created: newID'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request body or contact ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Contact not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiTokenAuth: []
      summary: Creates an Other Date associated with a contact
      tags:
      - contacts
  /api/v1/contacts/{cid}/other-dates/{oid}:
    delete:
      consumes:
      - application/json
      description: Removes an Other Date with a contact using HTTP DELETE
      parameters:
      - description: Contact ID
        in: path
        name: cid
        required: true
        type: integer
      - description: Other Date ID
        in: path
        name: oid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: deleted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request body or contact ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Contact not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiTokenAuth: []
      summary: Removes an Other Date associated with a contact
      tags:
      - contacts
...
  /api/v1/other-dates/{oid}:
    patch:
      consumes:
      - application/json
      description: Update specific fields of a phone using HTTP PATCH. Only provided
        fields will be updated.
      parameters:
      - description: Other Date ID
        in: path
        name: oid
        required: true
        type: integer
      - description: Other Date fields to update
        in: body
        name: contact
        required: true
        schema:
          $ref: '#/definitions/models.ContactDateJSONPatch'
      produces:
      - application/json
      responses:
        "200":
          description: Updated other date
          schema:
            items:
              $ref: '#/definitions/models.OtherDate'
            type: array
        "400":
          description: Invalid request body or contact ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Contact not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiTokenAuth: []
      summary: Update an other_date for a contact
      tags:
      - contacts
...
  /api/v1/settings/labels:
    post:
      consumes:
      - application/json
      description: Get birthdays, anniversaries, and other important dates coming
        up for all contacts
      parameters:
      - description: label fields
        in: body
        name: label
        required: true
        schema:
          $ref: '#/definitions/models.ContactLabelJSONPost'
      produces:
      - application/json
      responses:
        "200":
          description: id of label
          schema:
            items:
              additionalProperties:
                type: integer
              type: object
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiTokenAuth: []
      summary: Get upcoming events
      tags:
      - labels
  /api/v1/settings/labels/{lid}:
    delete:
      consumes:
      - application/json
      description: Removes a custom contact label using HTTP DELETE
      parameters:
      - description: Label ID
        in: path
        name: lid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: deleted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request body or Label ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Label not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiTokenAuth: []
      summary: Removes a custom contact label
      tags:
      - labels
...
securityDefinitions:
  ApiTokenAuth:
    description: API token for authentication
    in: header
    name: session
    type: apiKey
swagger: "2.0"